+++ ksp_plugin/ksp_plugin.vcxproj.filters
@@ -65,6 +65,9 @@
     <ClInclude Include="planetarium.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="iterators.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="interface.cpp">
